requirements.txt
mcp>=1.6.0
PyMySQL>=1.1.1
python-dotenv>=1.0.1
